requirements.txt
# Libraries

requests
Pillow